add exception handling

See merge request CJBaquiran/e-commerce-api!5

+++ src/main/kotlin/com/ecommerce/ecommerce/usecase/DeleteOneProductData.kt
@@ -1,6 +1,7 @@
 package com.ecommerce.ecommerce.usecase
 
 import com.ecommerce.ecommerce.domain.ProductRepository
+import com.ecommerce.ecommerce.exception.ProductServiceException
 import org.springframework.stereotype.Service
 import javax.transaction.Transactional
 
@@ -8,6 +9,9 @@ import javax.transaction.Transactional
 @Transactional
 class DeleteOneProductData(val productRepository: ProductRepository) {
     fun deleteOne(id: Long) {
+        if(productRepository.findById(id).id != id)
+            throw ProductServiceException("No Products to display")
+
         return productRepository.deleteById(id)
     }
 }
